+++ election-server/src/main/java/cropx/election/db/repository/CampaignRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import cropx.election.db.entity.Campaign;
 
-public interface CampaignRepository extends JpaRepository<Campaign, Long> {
+public interface CampaignRepository extends JpaRepository<Campaign, Integer> {
 
 }
